pr 14

+++ controllers/cards.js
@@ -1,8 +1,10 @@
 const Card = require('../models/card');
-const { checkId, checkErrors } = require('../utills/checks');
+const NotFoundError = require('../erorrs/not-found-error');
+const BadRequestError = require('../erorrs/bad-request-error');
+const ForbiddenError = require('../erorrs/forbidden-error');
 
-// POST
-const createCard = (req, res) => {
+// +
+const createCard = (req, res, next) => {
   const { _id } = req.user;
   const { name, link } = req.body;
   Card.create({ name, link, owner: _id })
@@ -10,36 +12,46 @@ const createCard = (req, res) => {
       res.send(card);
     })
     .catch((err) => {
-      checkErrors(err, res);
+      if (err.name === 'ValidationError') {
+        next(new BadRequestError('Переданы некорректные данные'));
+      }
+      next(err);
     });
 };
 
 // GET
-const getAllCards = (req, res) => {
+const getAllCards = (req, res, next) => {
   Card.find({})
     .populate('owner')
     .then((cards) => {
       res.send(cards);
     })
-    .catch((err) => {
-      checkErrors(err, res);
-    });
+    .catch(next);
 };
 
-// DELETE
-const deleteCard = (req, res) => {
+// +
+const deleteCard = (req, res, next) => {
+  const { _id } = req.user;
   const { cardId } = req.params;
-  Card.findByIdAndRemove(cardId)
+  Card.findById(cardId)
     .then((card) => {
-      checkId(card, res);
+      if (!card) {
+        throw new NotFoundError('Карточка или пользователь не найден');
+      }
+      if (card.owner.toString() !== _id) {
+        throw new ForbiddenError('Нет соответствующих прав');
+      }
+      Card.findByIdAndRemove(cardId)
+        .then((cardToDel) => {
+          res.send(cardToDel);
+        })
+        .catch(next);
     })
-    .catch((err) => {
-      checkErrors(err, res);
-    });
+    .catch(next);
 };
 
-// PUT
-const likeCard = (req, res) => {
+// +
+const likeCard = (req, res, next) => {
   const { _id } = req.user;
   const { cardId } = req.params;
   Card.findByIdAndUpdate(
@@ -48,15 +60,15 @@ const likeCard = (req, res) => {
     { new: true },
   )
     .then((card) => {
-      checkId(card, res);
+      if (!card) {
+        throw new NotFoundError('Карточка или пользователь не найден');
+      }
     })
-    .catch((err) => {
-      checkErrors(err, res);
-    });
+    .catch(next);
 };
 
-// PATCH
-const unlikeCard = (req, res) => {
+// +
+const unlikeCard = (req, res, next) => {
   const { _id } = req.user;
   const { cardId } = req.params;
   Card.findByIdAndUpdate(
@@ -65,11 +77,11 @@ const unlikeCard = (req, res) => {
     { new: true },
   )
     .then((card) => {
-      checkId(card, res);
+      if (!card) {
+        throw new NotFoundError('Карточка или пользователь не найден');
+      }
     })
-    .catch((err) => {
-      checkErrors(err, res);
-    });
+    .catch(next);
 };
 
 module.exports = {
